pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emergent-creativity"
version = "0.1.0"
description = "A 3D simulated apartment with a RL-driven neural-network tenant exploring emergent creativity."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["reinforcement-learning", "neural-network", "simulation", "3d", "physics"]
dependencies = [
    "pybullet>=3.2.5",
    "gymnasium>=0.29.1",
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "stable-baselines3>=2.2.1",
    "pygame>=2.5.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "tensorboard>=2.14.0",
    "scipy>=1.11.0",
]

[project.scripts]
ec-train = "scripts.train:main"
ec-view  = "scripts.view:main"

[project.optional-dependencies]
dev = ["pytest>=7.4", "pytest-cov>=4.1"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
